[Feature43] compose관련 패키지 구조 정리
- ui/ : Compose관련 UI코드
  - screen/ : 화면 단위 UI
  - component/ : 재사용 가능한 UI
