+++ app/src/main/java/com/example/objectdetection/ui/screen/DetailScreen.kt
@@ -1,0 +1,8 @@
+package com.example.objectdetection.ui.screen
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun DetailScreen(photoUrl: String, description: String) {
+
+}
